requirements.txt
Flask
gunicorn
transformers
torch
google-generativeai
requests
